requirements.txt
pysimdjson
python-telegram-bot
fastapi